+++ src/components/Background/MainBackground/MainBackground.tsx
@@ -1,0 +1,16 @@
+import * as S from './styles'
+
+const Background = (): JSX.Element => {
+  return (
+    <>
+      <S.LeftTopYellow/>
+      <S.LeftTopGreen/>
+      <S.RightTopGreen/>
+      <S.RightTopYellow/>
+      <S.GreenPolygon/>
+      <S.YellowHalf/>
+    </>
+  )
+}
+
+export default Background;
