+++ src/container/Main/index.tsx
@@ -1,0 +1,11 @@
+import Background from "../../components/Background/MainBackground/MainBackground";
+
+const MainContainer = (): JSX.Element => {
+  return (
+    <>
+      <Background />
+    </>
+  );
+};
+
+export default MainContainer;
